Move analytics and adapters to services folder under src
